requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
fuzzywuzzy==0.18.0
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
Werkzeug==3.0.3
zipp==3.19.2
